refactor(tests): move exported functions and variables to an additional file to reduce the number of tests.
