+++ tests/unit/components/LControl.test.ts
@@ -4,21 +4,14 @@ import LControl from '../../../src/components/LControl.vue'
 import { RegisterControlInjection } from '../../../src/types/injectionKeys'
 import { Control } from 'leaflet'
 import { testRemoveOnUnmount } from './helper/tests'
-import {
-    componentProps,
-    testComponentPropBindings,
-    testPropsBindingToLeaflet
-} from './helper/propsBindingTests'
+import { testComponentPropBindings, testPropsBindingToLeaflet } from './helper/propsBindingTests'
 import { testEmitsReady } from './helper/emitTests'
 import { mockRegisterControl, testControlRegistration } from './helper/injectionsTests'
 import { mergeReactiveProps } from './helper/props'
-import { createWrapper as createMapWrapper } from './LMap.test'
+import { createMapWrapper } from './wrapper/LMap'
+import { controlAbstractProps } from './wrapper/LControl'
 
-export const controlAbstractProps = mergeReactiveProps(componentProps, {
-    position: 'bottomleft'
-})
-
-export const controlProps = mergeReactiveProps(controlAbstractProps, {})
+const controlProps = mergeReactiveProps(controlAbstractProps, {})
 
 const createWrapper = async (props = {}) => {
     const wrapper = shallowMount(LControl, {
@@ -62,6 +55,7 @@ const testCorrectInitialisation = (getWrapper: () => Promise<VueWrapper<any>>) =
         const wrapper = await createMapWrapper({}, { default: LControl })
         const lControl = wrapper.findComponent(LControl)
         expect(lControl.vm.leafletObject).toBeDefined()
+        // @ts-expect-error _map is private so not in the types
         expect((lControl.vm.leafletObject as Control)._map).toBeDefined()
     })
 }
