+++ tests/unit/components/LImageOverlay.test.ts
@@ -2,27 +2,19 @@ import { flushPromises, shallowMount, type VueWrapper } from '@vue/test-utils'
 import { describe, expect, it } from 'vitest'
 import { AddLayerInjection, RemoveLayerInjection } from '../../../src/types/injectionKeys'
 import { testRemoveLayerOnUnmount } from './helper/tests'
-import {
-    layerProps,
-    testComponentPropBindings,
-    testPropsBindingToLeaflet
-} from './helper/propsBindingTests'
+import { testComponentPropBindings, testPropsBindingToLeaflet } from './helper/propsBindingTests'
 import { testEmitsReady } from './helper/emitTests'
 import { mockAddLayer, mockRemoveLayer, testAddLayer } from './helper/injectionsTests'
-import { ImageOverlay, LatLngBounds, type LatLngBoundsLiteral } from 'leaflet'
+import { ImageOverlay, type LatLngBoundsLiteral } from 'leaflet'
 import LImageOverlay from '../../../src/components/LImageOverlay.vue'
 import { mergeReactiveProps } from './helper/props'
-
-export const imageOverlayAbstractProps = mergeReactiveProps(layerProps, {
-    opacity: 0.5,
-    zIndex: 50,
-    bounds: new LatLngBounds([0, 0], [50, 50])
-})
+import { imageOverlayAbstractProps } from './wrapper/LImageOverlay'
 
 const imageOverlayProps = mergeReactiveProps(imageOverlayAbstractProps, {
     url: 'replace.jpg',
     expecting: {
         url: (leafletObject: ImageOverlay) => {
+            // @ts-expect-error _url is private so not in the types
             expect(leafletObject._url).toBe(imageOverlayProps.url)
         }
     }
