+++ tests/unit/components/LLayerGroup.test.ts
@@ -5,17 +5,11 @@ import LLayerGroup from '../../../src/components/LLayerGroup.vue'
 import LTileLayer from '../../../src/components/LTileLayer.vue'
 import { AddLayerInjection, RemoveLayerInjection } from '../../../src/types/injectionKeys'
 import { testRemoveLayerOnUnmount } from './helper/tests'
-import {
-    layerProps,
-    testComponentPropBindings,
-    testPropsBindingToLeaflet
-} from './helper/propsBindingTests'
+import { testComponentPropBindings, testPropsBindingToLeaflet } from './helper/propsBindingTests'
 import { testEmitsReady } from './helper/emitTests'
 import { mockAddLayer, mockRemoveLayer, testAddLayer } from './helper/injectionsTests'
 import { LayerGroup } from 'leaflet'
-import { mergeReactiveProps } from './helper/props'
-
-export const layerGroupProps = mergeReactiveProps(layerProps, {})
+import { layerGroupProps } from './wrapper/LLayerGroup'
 
 const createWrapper = async (props = {}, slots = {}) => {
     const wrapper = mount(LLayerGroup, {
