+++ tests/unit/components/wrapper/LControl.ts
@@ -1,0 +1,6 @@
+import { mergeReactiveProps } from '../helper/props'
+import { componentProps } from '../helper/propsBindingTests'
+
+export const controlAbstractProps = mergeReactiveProps(componentProps, {
+    position: 'bottomleft'
+})
