+++ tests/unit/components/wrapper/LLayerGroup.ts
@@ -1,0 +1,4 @@
+import { mergeReactiveProps } from '../helper/props'
+import { layerProps } from '../helper/propsBindingTests'
+
+export const layerGroupProps = mergeReactiveProps(layerProps, {})
